v10.0.21_prep

+++ testUpdate.sh
@@ -33,6 +33,41 @@ mainGitBlob=""
 versionPrepID=$(tr -dc A-Za-z0-9 </dev/urandom | head -c 25; echo)
 versionID=$(tr -dc A-Za-z0-9 </dev/urandom | head -c 25; echo)
 
+skipMainRepoCheck=0
+noRelease=0
+noServerUpdate=0
+noCleanup=0
+
+# parse args
+while [[ $# -gt 0 ]]; do
+	case $1 in
+		-C|--no-cleanup)
+			noCleanup=1
+			shift
+			;;
+		-S|--no-server-update)
+			noServerUpdate=1
+			shift
+			;;
+		-R|--no-release)
+			noRelease=1
+			shift
+			;;
+		-S|--skip-main-repo-check)
+			skipMainRepoCheck=1
+			shift
+			;;
+		-v|--version)
+			echo "v${version}"
+			shift
+			;;
+		-*|--*)
+			echo "Unknown option $1"
+			exit 1
+			;;
+	esac
+done
+
 ### update ###
 echo
 echo "### push prep to git ###"
@@ -77,19 +112,24 @@ cd ${tmpReleaseFileLocation}/${packName}_$packVersion
 echo PreLaunchCommand="\$INST_JAVA" -jar packwiz-installer-bootstrap.jar -s client ${packURL}/${mainGitBlob}/packwiz/pack.toml >> instance.cfg
 cd $workingDir
 
-
-./tools/createGithubRelease.sh \
-	--upstream "${packAPIURL}/releases" \
-	--tag "$packVersion" \
-	--name "v$packVersion" \
-	--description "$changelog" \
-	--branch "$branch" \
-	--token "$githubTokenPath" \
-	--release-folder "${tmpReleaseFileLocation}" \
-	--release-files-only
-
-rm -r ${tmpReleaseFileLocation}
-
+if [[ $noRelease == 1 ]]; then
+	echo "Skipping github release"
+else
+	./tools/createGithubRelease.sh \
+		--upstream "${packAPIURL}/releases" \
+		--tag "$packVersion" \
+		--name "v$packVersion" \
+		--description "$changelog" \
+		--branch "$branch" \
+		--token "$githubTokenPath" \
+		--release-folder "${tmpReleaseFileLocation}" \
+		--release-files-only
+end
+if [[ $noCleanup == 1 ]]; then
+	echo "Skipping release dir cleanup"
+else
+	rm -r ${tmpReleaseFileLocation}
+fi
 
 wait() {
 	while [[ $(wget -qO- $2) != $3 ]]; do	
@@ -98,19 +138,28 @@ wait() {
 	done
 }
 wait "Prep repository not updated yet. Wait another 30 seconds" ${packURL}/${prepGitBlob}/.versionID $versionPrepID
-wait "Main repository not updated yet. Wait another 30 seconds" ${packURL}/${branch}/.versionID $versionID
+wait "Main repository not updated yet. Wait another 30 seconds" ${packURL}/${mainGitBlob}/.versionID $versionID
+if [[ $skipMainRepoCheck == 1 ]]; then
+	echo "Skipping main branch check"
+else
+	wait "$branch branch not updated yet. Wait another 30 seconds" ${packURL}/${branch}/.versionID $versionID
+fi
 
 : '
-echo
-echo "### Update server ###"
-./tools/updateServer.sh \
-   --server-id $serverID \
-   --pterodactyl-url $pterodactylURL \
-   --pack-url "${packURL}/${branch}/packwiz/pack.toml" \
-   --token $pterodactylTokenPath \
-   --ssh-args="-i $sshKeyPath $sshTarget $sshArgs" \
-   --update-script-args="$updateScriptArgs" \
-	--no-backup \
-   $*
+if [[ $noServerUpdate == 1 ]]; then
+	echo "Skipping server update"
+else
+	echo
+	echo "### Update server ###"
+	./tools/updateServer.sh \
+	   --server-id $serverID \
+	   --pterodactyl-url $pterodactylURL \
+	   --pack-url "${packURL}/${branch}/packwiz/pack.toml" \
+	   --token $pterodactylTokenPath \
+	   --ssh-args="-i $sshKeyPath $sshTarget $sshArgs" \
+	   --update-script-args="$updateScriptArgs" \
+		--no-backup \
+	   $*
+fi
 	
 '
